requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
playwright==1.44.0
httpx==0.27.0
google-generativeai==0.5.4
pydantic==2.7.4
